Separating header and cpp
